backend/pyproject.toml
[project]
name = "tsmc-server"
version = "0.2.0"
description = "Self-hosted FastAPI backend and Linux service CLI for TSMC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.21.0",
  "fastapi>=0.115.0",
  "greenlet>=3.2.4",
  "httpx>=0.28.1",
  "orjson>=3.10.0",
  "python-dotenv>=1.1.1",
  "pwdlib[argon2]>=0.3.0",
  "pydantic-settings>=2.10.1",
  "sqlalchemy>=2.0.44",
  "uvicorn[standard]>=0.35.0",
  "uvloop>=0.22.1; platform_system != 'Windows'",
]

[project.scripts]
dev = "app.dev:main"
tsmc = "app.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.uv]
package = true
